Split stackwalking out of the minidump-processor crate.
The new minidump-walk-stack crate handles it and has far fewer
dependencies.

Closes #816.
