+++ minidump-processor/Cargo.toml
@@ -13,12 +13,15 @@ edition = "2018"
 travis-ci = { repository = "rust-minidump/rust-minidump" }
 
 [features]
+# Allows retrieval of symbols from debug info.
+debuginfo = ["minidump-walk-stack/debuginfo"]
 default = ["disasm_amd64"]
 # Allow disassembling of AMD64 instructions
 disasm_amd64 = ["yaxpeax-x86"]
 # Allows retrieval of symbol files via HTTP
-http = ["breakpad-symbols/http"]
+http = ["minidump-walk-stack/http"]
 # Allows construction of symbol files from native binaries
+# TODO why is http implied by dump_syms?
 dump_syms = ["breakpad-symbols/dump_syms", "http"]
 # Allows retrieval of CABed symbol files from mozilla servers
 mozilla_cab_symbols = ["breakpad-symbols/mozilla_cab_symbols"]
@@ -27,15 +30,14 @@ mozilla_cab_symbols = ["breakpad-symbols/mozilla_cab_symbols"]
 async-trait = "0.1.51"
 breakpad-symbols = { version = "0.16.0", path = "../breakpad-symbols" }
 debugid = "0.8.0"
-cachemap2 = "0.2.0"
 futures-util = "0.3.25"
 memmap2 = "0.5.7"
 minidump = { version = "0.16.0", path = "../minidump" }
 minidump-common = { version = "0.16.0", path = "../minidump-common" }
+minidump-walk-stack = { version = "0.16.0", path = "../minidump-walk-stack" }
 scroll = "0.11.0"
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
-symbolic = { version = "12", features = ["cfi", "debuginfo"] }
 thiserror = "1.0.37"
 tracing = { version = "0.1.34", features = ["log"] }
 yaxpeax-x86 = { version = "1.1", optional = true }
