+++ minidump-processor/src/arg_recovery.rs
@@ -1,5 +1,5 @@
-use crate::{CallStack, CallingConvention, FunctionArg, FunctionArgs};
 use minidump::{CpuContext, MinidumpRawContext, UnifiedMemory};
+use minidump_walk_stack::{CallStack, CallingConvention, FunctionArg, FunctionArgs};
 
 // # Recovering x86 function arguments
 //
