+++ minidump-stackwalk/Cargo.toml
@@ -25,9 +25,8 @@ clap = { version = "4.0.18", features = ["cargo", "wrap_help", "derive"] }
 indicatif = "0.17.0"
 minidump = { version = "0.16.0", path = "../minidump" }
 minidump-common = { version = "0.16.0", path = "../minidump-common" }
-minidump-processor = { version = "0.16.0", path = "../minidump-processor", features = [
-    "http",
-] }
+minidump-walk-stack = { version = "0.16.0", path = "../minidump-walk-stack", features = ["debuginfo", "http"] }
+minidump-processor = { version = "0.16.0", path = "../minidump-processor" }
 tokio = { version = "1.12.0", features = ["full"] }
 tracing = { version = "0.1.34", features = ["log"] }
 tracing-subscriber = "0.3.14"
