+++ minidump-walk-stack/Cargo.toml
@@ -1,0 +1,38 @@
+[package]
+name = "minidump-walk-stack"
+description = "A library for producing stack traces from minidump files."
+version = "0.16.0"
+authors = ["Alex Franchuk <afranchuk@mozilla.com>"]
+license = "MIT"
+readme = "README.md"
+homepage = "https://github.com/rust-minidump/rust-minidump"
+repository = "https://github.com/rust-minidump/rust-minidump"
+edition = "2018"
+
+[badges]
+travis-ci = { repository = "rust-minidump/rust-minidump" }
+
+[features]
+# Allows retrieval of symbols from debug info.
+debuginfo = ["cachemap2", "futures-util", "memmap2", "symbolic"]
+# Allows retrieval of symbol files via HTTP
+http = ["breakpad-symbols/http"]
+
+[dependencies]
+async-trait = "0.1.51"
+breakpad-symbols = { version = "0.16.0", path = "../breakpad-symbols" }
+cachemap2 = { version = "0.2.0", optional = true }
+futures-util = { version = "0.3.25", optional = true }
+memmap2 = { version = "0.5.7", optional = true }
+minidump = { version = "0.16.0", path = "../minidump" }
+minidump-common = { version = "0.16.0", path = "../minidump-common" }
+scroll = "0.11.0"
+symbolic = { version = "12", features = ["cfi", "debuginfo"], optional = true }
+tracing = { version = "0.1.34", features = ["log"] }
+
+[dev-dependencies]
+test-assembler = "0.1.6"
+tokio = { version = "1.12.0", features = ["full"] }
+
+[package.metadata.docs.rs]
+all-features = true
