+++ minidump-walk-stack/src/x86.rs
@@ -6,28 +6,23 @@
 // so that it can be copied between the two with minimal changes. It's not
 // worth the effort to *actually* unify the implementations.
 
-use crate::process_state::{FrameTrust, StackFrame};
-use crate::stackwalker::unwind::Unwind;
-use crate::stackwalker::CfiStackWalker;
-use crate::{SymbolProvider, SystemInfo};
-use minidump::format::CONTEXT_AMD64;
-use minidump::system_info::Os;
+use super::impl_prelude::*;
+use minidump::format::CONTEXT_X86;
 use minidump::{
     MinidumpContext, MinidumpContextValidity, MinidumpModuleList, MinidumpRawContext, UnifiedMemory,
 };
 use std::collections::HashSet;
 use tracing::trace;
 
-type Pointer = u64;
-const POINTER_WIDTH: Pointer = 8;
-const INSTRUCTION_REGISTER: &str = "rip";
-const STACK_POINTER_REGISTER: &str = "rsp";
-const FRAME_POINTER_REGISTER: &str = "rbp";
-// FIXME: rdi and rsi are also preserved on windows (but not in sysv) -- we should handle that?
-const CALLEE_SAVED_REGS: &[&str] = &["rbx", "rbp", "r12", "r13", "r14", "r15"];
+type Pointer = u32;
+const POINTER_WIDTH: Pointer = 4;
+const INSTRUCTION_REGISTER: &str = "eip";
+const STACK_POINTER_REGISTER: &str = "esp";
+const FRAME_POINTER_REGISTER: &str = "ebp";
+const CALLEE_SAVED_REGS: &[&str] = &["ebp", "ebx", "edi", "esi"];
 
 async fn get_caller_by_cfi<P>(
-    ctx: &CONTEXT_AMD64,
+    ctx: &CONTEXT_X86,
     callee: &StackFrame,
     grand_callee: Option<&StackFrame>,
     stack_memory: UnifiedMemory<'_, '_>,
@@ -71,11 +66,11 @@ where
     symbol_provider
         .walk_frame(module, &mut stack_walker)
         .await?;
-    let caller_ip = stack_walker.caller_ctx.rip;
-    let caller_sp = stack_walker.caller_ctx.rsp;
+    let caller_ip = stack_walker.caller_ctx.eip;
+    let caller_sp = stack_walker.caller_ctx.esp;
 
     trace!(
-        "cfi evaluation was successful -- caller_ip: 0x{:016x}, caller_sp: 0x{:016x}",
+        "cfi evaluation was successful -- caller_ip: 0x{:08x}, caller_sp: 0x{:08x}",
         caller_ip,
         caller_sp,
     );
@@ -85,10 +80,23 @@ where
     // values are correct. I Don't Like This, but it's what breakpad does and
     // we should start with a baseline of parity.
 
+    // FIXME?: breakpad is actually a little weary of the output of STACK WIN
+    // cfi, and does check that instruction_seems_valid() for eip. However,
+    // it doesn't immediately discard the results. It tentatively tries to
+    // scan, and then if that doesn't return anything compelling, it just goes
+    // forward with whatever STACK WIN came up with.
+    //
+    // The current layering of this code means that we don't actually know what
+    // kind of cfi was used here, and the code that *does* can't do scanning.
+    // For now let's just trust the results unconditionally. We can do something
+    // more hacky/robust if we find a compelling need to.
+    //
+    // It also has some weird scanning to try to adjust the computed bp?
+
     trace!("cfi result seems valid");
 
     let context = MinidumpContext {
-        raw: MinidumpRawContext::Amd64(stack_walker.caller_ctx),
+        raw: MinidumpRawContext::X86(stack_walker.caller_ctx),
         valid: MinidumpContextValidity::Some(stack_walker.caller_validity),
     };
     Some(StackFrame::from_context(context, FrameTrust::CallFrameInfo))
@@ -106,44 +114,27 @@ fn callee_forwarded_regs(valid: &MinidumpContextValidity) -> HashSet<&'static st
 }
 
 fn get_caller_by_frame_pointer<P>(
-    ctx: &CONTEXT_AMD64,
+    ctx: &CONTEXT_X86,
     callee: &StackFrame,
     stack_memory: UnifiedMemory<'_, '_>,
     _modules: &MinidumpModuleList,
-    system_info: &SystemInfo,
     _symbol_provider: &P,
 ) -> Option<StackFrame>
 where
     P: SymbolProvider + Sync,
 {
-    // On Windows x64, frame-pointer unwinding purely with the data on the stack
-    // is not possible, as proper unwinding requires access to `UNWIND_INFO`,
-    // because the frame pointer does not necessarily point to the end of the
-    // frame.
-    // In particular, the docs state that:
-    // > [The frame register] offset permits pointing the FP register into the
-    // > middle of the local stack allocation [...]
-    // https://docs.microsoft.com/en-us/cpp/build/exception-handling-x64
-    if system_info.os == Os::Windows {
-        return None;
-    }
-
     trace!("trying frame pointer");
     if let MinidumpContextValidity::Some(ref which) = callee.context.valid {
         if !which.contains(FRAME_POINTER_REGISTER) {
             return None;
         }
-        if !which.contains(STACK_POINTER_REGISTER) {
-            return None;
-        }
     }
 
-    let last_bp = ctx.rbp;
-    let last_sp = ctx.rsp;
-    // Assume that the standard %bp-using x64 calling convention is in
+    let last_bp = ctx.ebp;
+    // Assume that the standard %bp-using x86 calling convention is in
     // use.
     //
-    // The typical x64 calling convention, when frame pointers are present,
+    // The typical x86 calling convention, when frame pointers are present,
     // is for the calling procedure to use CALL, which pushes the return
     // address onto the stack and sets the instruction pointer (%ip) to
     // the entry point of the called routine.  The called routine then
@@ -156,71 +147,54 @@ where
     // calling procedure's stack pointer (%sp) is 2 pointers higher than the
     // value of the called procedure's frame pointer at the time the calling
     // procedure made the CALL: 1 pointer for the return address pushed by the
-    // CALL itself, and 1 pointer for the callee's` PUSH of the caller's frame
+    // CALL itself, and 1 pointer for the callee's PUSH of the caller's frame
     // pointer.
     //
     // %ip_new = *(%bp_old + ptr)
     // %bp_new = *(%bp_old)
     // %sp_new = %bp_old + ptr*2
 
-    if last_bp >= u64::MAX - POINTER_WIDTH * 2 {
+    if last_bp >= u32::MAX - POINTER_WIDTH * 2 {
         // Although this code generally works fine if the pointer math overflows,
         // debug builds will still panic, and this guard protects against it without
         // drowning the rest of the code in checked_add.
         return None;
     }
-    let caller_ip = stack_memory.get_memory_at_address(last_bp + POINTER_WIDTH)?;
-    let caller_bp = stack_memory.get_memory_at_address(last_bp)?;
+    let caller_ip = stack_memory.get_memory_at_address(last_bp as u64 + POINTER_WIDTH as u64)?;
+    let caller_bp = stack_memory.get_memory_at_address(last_bp as u64)?;
     let caller_sp = last_bp + POINTER_WIDTH * 2;
 
-    // If the recovered ip is not a canonical address it can't be
-    // the return address, so bp must not have been a frame pointer.
-
-    // Since we're assuming coherent frame pointers, check that the frame pointers
-    // and stack pointers are well-ordered.
-    if caller_sp <= last_bp || caller_bp < caller_sp {
-        trace!("rejecting frame pointer result for unreasonable frame pointer");
-        return None;
-    }
-    // Since we're assuming coherent frame pointers, check that the resulting
-    // frame pointer is still inside stack memory.
-    let _unused: Pointer = stack_memory.get_memory_at_address(caller_bp)?;
-    // Don't accept obviously wrong instruction pointers.
-    if is_non_canonical(caller_ip) {
-        trace!("rejecting frame pointer result for unreasonable instruction pointer");
-        return None;
-    }
-    // Don't accept obviously wrong stack pointers.
-    if !stack_seems_valid(caller_sp, last_sp, stack_memory) {
-        trace!("rejecting frame pointer result for unreasonable stack pointer");
-        return None;
-    }
+    // NOTE: minor divergence from x64 impl here: doing extra validation on the
+    // value of `caller_sp` and `caller_bp` here encourages the stack scanner
+    // to kick in and start outputting extra frames for `/testdata/test.dmp`.
+    // Since breakpad also doesn't output those frames, let's assume that's
+    // desirable.
 
     trace!(
-        "frame pointer seems valid -- caller_ip: 0x{:016x}, caller_sp: 0x{:016x}",
+        "frame pointer seems valid -- caller_ip: 0x{:08x}, caller_sp: 0x{:08x}",
         caller_ip,
         caller_sp,
     );
 
-    let caller_ctx = CONTEXT_AMD64 {
-        rip: caller_ip,
-        rsp: caller_sp,
-        rbp: caller_bp,
-        ..CONTEXT_AMD64::default()
+    let caller_ctx = CONTEXT_X86 {
+        eip: caller_ip,
+        esp: caller_sp,
+        ebp: caller_bp,
+        ..CONTEXT_X86::default()
     };
     let mut valid = HashSet::new();
     valid.insert(INSTRUCTION_REGISTER);
     valid.insert(STACK_POINTER_REGISTER);
     valid.insert(FRAME_POINTER_REGISTER);
     let context = MinidumpContext {
-        raw: MinidumpRawContext::Amd64(caller_ctx),
+        raw: MinidumpRawContext::X86(caller_ctx),
         valid: MinidumpContextValidity::Some(valid),
     };
     Some(StackFrame::from_context(context, FrameTrust::FramePointer))
 }
 
 async fn get_caller_by_scan<P>(
-    ctx: &CONTEXT_AMD64,
+    ctx: &CONTEXT_X86,
     callee: &StackFrame,
     stack_memory: UnifiedMemory<'_, '_>,
     modules: &MinidumpModuleList,
@@ -237,20 +211,20 @@ where
     // the current frame. The next frame is then assumed to end just before that
     // ip value.
     let last_bp = match callee.context.valid {
-        MinidumpContextValidity::All => Some(ctx.rbp),
+        MinidumpContextValidity::All => Some(ctx.ebp),
         MinidumpContextValidity::Some(ref which) => {
             if !which.contains(STACK_POINTER_REGISTER) {
                 trace!("cannot scan without stack pointer");
                 return None;
             }
             if which.contains(FRAME_POINTER_REGISTER) {
-                Some(ctx.rbp)
+                Some(ctx.ebp)
             } else {
                 None
             }
         }
     };
-    let last_sp = ctx.rsp;
+    let last_sp = ctx.esp;
 
     // Number of pointer-sized values to scan through in our search.
     let default_scan_range = 40;
@@ -266,7 +240,7 @@ where
 
     for i in 0..scan_range {
         let address_of_ip = last_sp.checked_add(i * POINTER_WIDTH)?;
-        let caller_ip = stack_memory.get_memory_at_address(address_of_ip)?;
+        let caller_ip = stack_memory.get_memory_at_address(address_of_ip as u64)?;
         if instruction_seems_valid(caller_ip, modules, symbol_provider).await {
             // ip is pushed by CALL, so sp is just address_of_ip + ptr
             let caller_sp = address_of_ip.checked_add(POINTER_WIDTH)?;
@@ -285,38 +259,38 @@ where
             // simple heuristics like "is a valid stack address".
             let mut caller_bp = None;
 
-            // This value was specifically computed for x86 frames (see the x86
-            // impl for details), but 128 KB is still an extremely generous
-            // frame size on x64.
+            // Max reasonable size for a single x86 frame is 128 KB.  This value is used in
+            // a heuristic for recovering of the EBP chain after a scan for return address.
+            // This value is based on a stack frame size histogram built for a set of
+            // popular third party libraries which suggests that 99.5% of all frames are
+            // smaller than 128 KB.
             const MAX_REASONABLE_GAP_BETWEEN_FRAMES: Pointer = 128 * 1024;
 
-            // NOTE: minor divergence from the x86 impl here: for whatever
-            // reason the x64 breakpad tests only work if we gate option (1) on
-            // having a valid `bp` that points next to address_of_ip already.
-            // It's unclear why, perhaps the test is buggy, but for now we
-            // preserve that behaviour.
-            if let Some(last_bp) = last_bp {
-                // If we're on the first iteration of the scan, there can't possibly be a
-                // frame pointer, because the entire stack frame is taken up by the return
-                // pointer. And if we're not on the first iteration, then the last iteration
-                // already loaded the location we expect the frame pointer to be in, so we can
-                // unconditionally load it here.
-                if i > 0 {
-                    let address_of_bp = address_of_ip - POINTER_WIDTH;
-                    // Can assume this resolves because we already walked over it when
-                    // checking address_of_ip values.
-                    let bp = stack_memory.get_memory_at_address(address_of_bp)?;
-                    if last_bp == address_of_bp
-                        && bp > address_of_ip
-                        && bp - address_of_bp <= MAX_REASONABLE_GAP_BETWEEN_FRAMES
+            // If we're on the first iteration of the scan, there can't possibly be a frame pointer,
+            // because the entire stack frame is taken up by the return pointer. And if we're
+            // not on the first iteration, then the last iteration already loaded the location
+            // we expect the frame pointer to be in, so we can unconditionally load it here.
+            if i > 0 {
+                let address_of_bp = address_of_ip - POINTER_WIDTH;
+                let bp = stack_memory.get_memory_at_address(address_of_bp as u64)?;
+
+                if bp > address_of_ip && bp - address_of_bp <= MAX_REASONABLE_GAP_BETWEEN_FRAMES {
+                    // Sanity check that resulting bp is still inside stack memory.
+                    if stack_memory
+                        .get_memory_at_address::<Pointer>(bp as u64)
+                        .is_some()
                     {
-                        // Final sanity check that resulting bp is still inside stack memory.
-                        if stack_memory.get_memory_at_address::<Pointer>(bp).is_some() {
-                            caller_bp = Some(bp);
+                        caller_bp = Some(bp);
+                    }
+                } else if let Some(last_bp) = last_bp {
+                    if last_bp >= caller_sp {
+                        // Sanity check that resulting bp is still inside stack memory.
+                        if stack_memory
+                            .get_memory_at_address::<Pointer>(last_bp as u64)
+                            .is_some()
+                        {
+                            caller_bp = Some(last_bp);
                         }
-                    } else if last_bp >= caller_sp {
-                        // Don't sanity check that the address is inside the stack? Hmm.
-                        caller_bp = Some(last_bp);
                     }
                 }
             }
@@ -327,11 +301,11 @@ where
                 caller_sp,
             );
 
-            let caller_ctx = CONTEXT_AMD64 {
-                rip: caller_ip,
-                rsp: caller_sp,
-                rbp: caller_bp.unwrap_or(0),
-                ..CONTEXT_AMD64::default()
+            let caller_ctx = CONTEXT_X86 {
+                eip: caller_ip,
+                esp: caller_sp,
+                ebp: caller_bp.unwrap_or(0),
+                ..CONTEXT_X86::default()
             };
             let mut valid = HashSet::new();
             valid.insert(INSTRUCTION_REGISTER);
@@ -340,7 +314,7 @@ where
                 valid.insert(FRAME_POINTER_REGISTER);
             }
             let context = MinidumpContext {
-                raw: MinidumpRawContext::Amd64(caller_ctx),
+                raw: MinidumpRawContext::X86(caller_ctx),
                 valid: MinidumpContextValidity::Some(valid),
             };
             return Some(StackFrame::from_context(context, FrameTrust::Scan));
@@ -381,13 +355,16 @@ async fn instruction_seems_valid<P>(
 where
     P: SymbolProvider + Sync,
 {
-    if is_non_canonical(instruction) || instruction == 0 {
+    if instruction == 0 {
         return false;
     }
 
-    super::instruction_seems_valid_by_symbols(instruction, modules, symbol_provider).await
+    super::instruction_seems_valid_by_symbols(instruction as u64, modules, symbol_provider).await
 }
 
+/*
+// x86 is currently hyper-permissive, so we don't use this,
+// but here it is in case we change our minds!
 fn stack_seems_valid(
     caller_sp: Pointer,
     callee_sp: Pointer,
@@ -400,35 +377,20 @@ fn stack_seems_valid(
 
     // The stack pointer should be in the stack
     stack_memory
-        .get_memory_at_address::<Pointer>(caller_sp)
+        .get_memory_at_address::<Pointer>(caller_sp as u64)
         .is_some()
 }
-
-fn is_non_canonical(ptr: Pointer) -> bool {
-    // x64 has the notion of a "canonical address", as a result of only 48 bits
-    // of a pointer actually being used, because this is all that a 4-level page
-    // table can support. A canonical address copies bit 47 to all the otherwise
-    // unused high bits. This creates two ranges where no valid pointers should
-    // ever exist.
-    //
-    // Note that as of this writing, 5-level page tables *do* exist, and when enabled
-    // 57 bits are used. However modern JS engines rely on only 48 bits being used
-    // to perform "NaN boxing" optimizations, so it's reasonable to assume
-    // by default that only 4-level page tables are used. (Even if enabled at
-    // the system level, Linux only exposes non-48-bit pointers to a process
-    // if that process explicitly opts in with a special operation.)
-    ptr > 0x7FFFFFFFFFFF && ptr < 0xFFFF800000000000
-}
+*/
 
 #[async_trait::async_trait]
-impl Unwind for CONTEXT_AMD64 {
+impl Unwind for CONTEXT_X86 {
     async fn get_caller_frame<P>(
         &self,
         callee: &StackFrame,
         grand_callee: Option<&StackFrame>,
         stack_memory: Option<UnifiedMemory<'_, '_>>,
         modules: &MinidumpModuleList,
-        system_info: &SystemInfo,
+        _system_info: &SystemInfo,
         syms: &P,
     ) -> Option<StackFrame>
     where
@@ -442,7 +404,7 @@ impl Unwind for CONTEXT_AMD64 {
             frame = get_caller_by_cfi(self, callee, grand_callee, stack, modules, syms).await;
         }
         if frame.is_none() {
-            frame = get_caller_by_frame_pointer(self, callee, stack, modules, system_info, syms);
+            frame = get_caller_by_frame_pointer(self, callee, stack, modules, syms);
         }
         if frame.is_none() {
             frame = get_caller_by_scan(self, callee, stack, modules, syms).await;
@@ -462,7 +424,7 @@ impl Unwind for CONTEXT_AMD64 {
         // If the new stack pointer is at a lower address than the old,
         // then that's clearly incorrect. Treat this as end-of-stack to
         // enforce progress and avoid infinite loops.
-        if frame.context.get_stack_pointer() <= self.rsp {
+        if frame.context.get_stack_pointer() <= self.esp as u64 {
             trace!("stack pointer went backwards, assuming unwind complete");
             return None;
         }
